requirements.txt
snap-plugin-lib-py
ruamel.yaml
git+https://github.com/michep/python-crsctl.git
